Added some Javadoc  for validate game also noticed possible validation mistake
There is a check for the number of cards in the players hands to make
sure that the players have 3 cards unless they are currently playing.
However shouldn't they have less than 3 cards if the game is a load from
an almost finished game? I put in a TODO so someone can doublecheck
this.